+++ R/radsl.R
@@ -87,12 +87,20 @@ radsl <- function(N = 400, # nolint
     ) %>%
       sample_fct(N, prob = c(.55, .23, .16, .05, .004, .003, .002, .002)),
     TRTSDTM = as.POSIXct(
-      sys_dtm + sample(seq(0, study_duration_secs), size = N, replace = TRUE),
+      sys_dtm + sample(seq(0, study_duration_secs/8), size = N, replace = TRUE),
       origin = "1970-01-01"
     ),
+    TRT01SDTM = TRTSDTM,
+    AP01SDTM = TRT01SDTM,
     RANDDT = as.Date(.data$TRTSDTM) - floor(stats::runif(N, min = 0, max = 5)),
-    st_posixn = as.numeric(.data$TRTSDTM),
-    TRTEDTM = as.POSIXct(.data$st_posixn + study_duration_secs, origin = "1970-01-01"),
+    st_posixn = as.numeric(.data$TRT01SDTM),
+    TRT01EDTM = as.POSIXct(.data$st_posixn + study_duration_secs/4, origin = "1970-01-01"),
+    AP01EDTM = TRT01EDTM,
+    st_posixn_2 = as.numeric(.data$TRT01EDTM),
+    TRT02SDTM = as.POSIXct(.data$st_posixn_2 + study_duration_secs/8, origin = "1970-01-01"),
+    AP02SDTM = TRT02SDTM,
+    st_posixn_3 = as.numeric(.data$TRT02SDTM),
+    TRTEDTM = as.POSIXct(.data$st_posixn_3 + study_duration_secs/2, origin = "1970-01-01"),
     STRATA1 = c("A", "B", "C") %>% sample_fct(N),
     STRATA2 = c("S1", "S2") %>% sample_fct(N),
     BMRKR1 = stats::rchisq(N, 6),
@@ -133,7 +141,10 @@ radsl <- function(N = 400, # nolint
       st_posixn >= quantile(st_posixn)[2] & st_posixn <= quantile(st_posixn)[3] ~ as.POSIXct(NA, origin = "1970-01-01"),
       TRUE ~ TRTEDTM
     )) %>%
-    dplyr::mutate(TRTEDTM = as.POSIXct(.data$TRTEDTM, origin = "1970-01-01")) %>%
+    dplyr::mutate(TRTEDTM = as.POSIXct(.data$TRTEDTM, origin = "1970-01-01"),
+                  AP02EDTM = TRTEDTM,
+                  TRT02EDTM = TRTEDTM
+                  ) %>%
     dplyr::select(-.data$TRTEDTM_discon)
 
   ADSL <- ADSL %>% # nolint
@@ -187,6 +198,13 @@ radsl <- function(N = 400, # nolint
       DTHCAT == "OTHER" ~ sample(x = l_dthcat_other$choices, size = N, replace = TRUE, prob = l_dthcat_other$prob),
       TRUE ~ as.character(NA)
     )) %>%
+    dplyr::mutate(ADTHAUT = dplyr::case_when(
+      DTHCAUS %in% c("ADVERSE EVENT", "DISEASE PROGRESSION") ~ "Yes",
+      DTHCAUS %in% c("UNKNOWN", "SUICIDE", "Post-study reporting of death") ~ sample(
+        x = c("Yes", "No"), size = N, replace = TRUE, prob = c(0.25, 0.75)
+      ),
+      TRUE ~ as.character(NA)
+    )) %>%
     # adding some random number of days post last treatment date so that death days from last trt admin
     # supports the LDDTHGR1 derivation below
     dplyr::mutate(DTHDT = dplyr::case_when(
@@ -202,7 +220,7 @@ radsl <- function(N = 400, # nolint
       DCSREAS == "DEATH" ~ DTHDT,
       TRUE ~ as.Date(.data$TRTEDTM) + floor(stats::runif(N, min = 10, max = 30))
     )) %>%
-    dplyr::select(-.data$st_posixn)
+    dplyr::select(-.data$st_posixn, -.data$st_posixn_2, -.data$st_posixn_3)
 
   # add random ETHNIC (Ethnicity)
   ADSL <- ADSL %>% # nolint